Resolved issue with duplicate tracks
When editing a track, the system checks if a track with new artist/title
already exists, and if it does, the content link is adjusted, while the
track remains unchanged, unless it has no content or relations with
other tracks, in which case it is removed.
Added user avatar rounding.